refactor: move inner.clipboard to clipboard
